task/#144 Refactor clan service
